스프링 시큐리티 / Jwt 리팩토링

+++ src/main/java/com/example/mini_project/dto/responseDto/BoardResponseDto.java
@@ -19,7 +19,7 @@ public class BoardResponseDto {
     private String content;
     private String writer;
     private String name;
-    private int heartNum;
+    private long heartNum;
     private List<CommentResponseDto> commentResponseDtoList;
     private LocalDateTime createdAt;
